Moved the Codecs enum into its own file in the codecs crate.
